+++ lib/ooxml_parser/pptx_parser/pptx_data/presentation/slide/slide/shapes_grouping/grouping_properties.rb
@@ -1,3 +1,4 @@
+require_relative 'groupoing_properties/transform_effect'
 module OoxmlParser
   class GroupingProperties
     attr_accessor :transform