+++ spec/password_protected_documents_spec.rb
@@ -1,5 +1,4 @@
-require 'rspec'
-require 'ooxml_parser'
+require 'spec_helper'
 
 describe 'Password protected documents' do
   it 'Password protected docx' do