+++ spec/spec_helper.rb
@@ -1,2 +1,6 @@
+require 'simplecov'
 require 'codeclimate-test-reporter'
+
 CodeClimate::TestReporter.start
+SimpleCov.start
+require 'ooxml_parser'