+++ spec/xlsx_examples/chart_types_spec.rb
@@ -1,5 +1,4 @@
-require 'rspec'
-require 'ooxml_parser'
+require 'spec_helper'
 
 describe 'Chart types' do
   it 'Chart Line 3d' do