+++ Changelog.md
@@ -11,6 +11,11 @@
 * Fix misplaced `dxa` and `emu` units of measurements and also fix calculation `dxa` unit
 * Redone parsing of `nary` in formulas
 
+### Refactor
+* Move parsing Columns inside class
+* Move parsing TableGrid inside class
+* Move parsing DocxShapeLinePath inside class
+
 ## 0.1.2 (2016-06-07)
 
 ### New features